% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram_outputs-data.R
\docType{data}
\name{nomogram_outputs}
\alias{nomogram_outputs}
\title{Nomogram outputs using the predicted probability of binary outcome}
\format{
A data frame with 16 rows and 1 column:
\describe{
\item{output}{A binary outcome with values from 0 to 1.}
}
}
\source{
Generated by a caret randomforest model using categorical predictors
for examples in this package.
}
\usage{
nomogram_outputs
}
\description{
An example of a data frame for \code{sample_output} argument in
\code{\link{create_nomogram}} function, must only include the predicted
probabilities for binary outcome.
}
\keyword{dataset}
