% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram_outputs4-data.R
\docType{data}
\name{nomogram_outputs4}
\alias{nomogram_outputs4}
\title{Nomogram outputs using the estimated value of numerical outcome}
\format{
A data frame with 80 rows and 1 column:
\describe{
\item{output}{A numerical outcome.}
}
}
\source{
Generated by a caret randomforest model using categorical and 1
numerical predictors for examples in this package.
}
\usage{
nomogram_outputs4
}
\description{
An example of a data frame for \code{sample_output} argument in
\code{\link{create_nomogram}} function, must only include the estimated
values for numerical outcome.
}
\keyword{dataset}
